requirements.txt
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
airportsdata==20250224
annotated-types==0.7.0
anyio==4.6.2.post1
astor==0.8.1
async-timeout==4.0.3
attrs==24.2.0
av==14.1.0
blake3==1.0.4
blis==1.1.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
clean-fid==0.1.35
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpathlib==0.20.0
cloudpickle==3.1.1
compressed-tensors==0.9.1
confection==0.1.5
contourpy==1.3.0
cupy-cuda12x==13.4.0
cycler==0.12.1
cymem==2.0.10
datasets==3.0.2
decord==0.6.0
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.0
email_validator==2.2.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
fastapi==0.115.11
fastapi-cli==0.0.7
fastrlock==0.8.3
filelock==3.16.1
flash_attn==2.7.4.post1
fonttools==4.54.1
freetype-py==2.5.1
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.3.1
gguf==0.10.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.1
idna==3.10
imageio==2.36.1
importlib_metadata==8.6.1
importlib_resources==6.4.5
iniconfig==2.0.0
interegular==0.3.3
Jinja2==3.1.6
jiter==0.6.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
lark==1.2.2
llvmlite==0.43.0
lm-format-enforcer==0.10.11
manifold3d==2.5.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mistral_common==1.5.3
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.11
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
openai==1.52.1
opencv-python-headless==4.11.0.86
outlines==0.1.11
outlines_core==0.1.26
packaging==24.1
pandas==2.2.3
partial-json-parser==0.2.1.1.post5
peft==0.15.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
preshed==3.0.9
prometheus-fastapi-instrumentator==7.0.2
prometheus_client==0.21.1
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pybind11==2.13.6
pycountry==24.6.1
pydantic==2.9.2
pydantic_core==2.23.4
pyglet==2.1.0
Pygments==2.18.0
PyOpenGL==3.1.0
pyparsing==3.2.0
pyrender==0.1.45
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.1
qwen-vl-utils==0.0.8
ray==2.40.0
referencing==0.36.2
regex==2024.9.11
requests==2.32.3
rich==13.9.3
rich-toolkit==0.13.2
rpds-py==0.23.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.17.0
setproctitle==1.3.3
shapely==2.0.6
shellingham==1.5.4
shtab==1.7.1
six==1.16.0
smart-open==7.1.0
smmap==5.0.1
sniffio==1.3.1
srsly==2.5.0
starlette==0.46.0
sympy==1.13.1
thinc==8.3.3
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.66.5
transformers==4.49.0
triangle==20230923
trimesh==4.5.1
triton==3.1.0
trl==0.15.2
typer==0.15.1
typing_extensions==4.12.2
tyro==0.8.14
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
vllm==0.7.3
wandb==0.18.5
wasabi==1.1.3
watchfiles==1.0.4
wcwidth==0.2.13
weasel==0.4.1
websockets==15.0.1
wrapt==1.17.0
xformers==0.0.28.post3
xgrammar==0.1.11
xxhash==3.5.0
yarl==1.16.0
zipp==3.20.2
